requirements.txt
# Core AI + LLMs
openai
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-classic

# Vector store / database
supabase

# YouTube + audio
yt-dlp
youtube-transcript-api
pydub


# PDF export
fpdf2
reportlab

# Environment Variables
python-dotenv

# FastAPI
fastapi
uvicorn[standard]